# data_to_wide: improve speed