Prepare the benchmark:
``` r
library(tidyr)
library(datawizard)

# Make fake "big" data
df <- fish_encounters 
df$fish <- as.numeric(df$fish)

all_df <- list()

for (i in 1:200) {
  all_df[[i]] <- df 
  all_df[[i]]$fish <- all_df[[i]]$fish + 20*i
}

all <- tibble::as_tibble(data.table::rbindlist(all_df))
```

Current results:
```r
# Benchmark
bench::mark(
  datawizard = all %>% reshape_wider(names_from = "station", values_from = "seen")
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 datawizard    478ms    485ms      2.06      26MB     20.6
```
New results:
``` r
bench::mark(
  datawizard = all %>% reshape_wider(names_from = "station", values_from = "seen")
)
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 datawizard   95.9ms   96.9ms      10.3    20.1MB     20.6
```
For information, `tidyr::pivot_wider()` is still far ahead:
``` r
bench::mark(
  tidyr = all %>% pivot_wider(names_from = station, values_from = seen)
)
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 tidyr        6.26ms   6.91ms      140.    3.48MB     17.5
```