@strengejacke 2022-09-05
very nice! once tests pass, we can merge this one. And yes, we can't compare our benchmarks with tidyr/dplyr. I think we're slower by some magnitudes due to relying on probably slower base R functions. For daily work with "normal" data sets it's ok, but for larger datasets, you may need to switch to dolyr/tidyr or even data.table/dtplyr.