@etiennebacher 2022-09-05
Yes especially because these pivot functions are partly implemented in C in tidyr, so I don't expect to compete with them. The benchmark was just to create some motivation to continue optimizing this ;)