@IndrajeetPatil 2022-09-05
Thanks, @etiennebacher! 

Improving performance is going to be the top most priority once we release the stable (1.0) versions of R packages. But, until then, we should definitely try to pick some low-hanging fruits that help with this.